Avoid snprintf when loading modules
Clang warned about out of bounds writes here on some platforms, though I think
it's a false positive.  In any case, it's hard to tell because this "calculate
then snprintf and hope it fits" pattern is error-prone.  Replace it with one
that is more verbose, but also more explicit and regular with no room for
misinterpretation.